requirements.txt
chainlit
python-dotenv
google-genai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
docx2txt
tqdm
numpy
setuptools<82
typing_extensions>=4.12.0
pydantic>=2.0